add split6
